+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## [v3.0.0] (2025-08-13)
+
+*Breaking change*: For React Native apps integrating with the native Android/iOS Performance SDKs, the minimum required versions are now: 
+- Bugsnag Android Performance [v2.0.0](https://github.com/bugsnag/bugsnag-android-performance/releases/tag/v2.0.0) 
+- Bugsnag Cocoa Performance [v1.14.0](https://github.com/bugsnag/bugsnag-cocoa-performance/releases/tag/v1.14.0)
+
+Upgrading these versions is required for continued compatibility with the React Native integration. Please refer to the linked release notes for details on changes in the native SDKs.
+
+### Added
+
+- (react-native) Added `BugsnagJavascriptSpansPlugin` to `@bugsnag/plugin-react-native-span-access` to support accessing JavaScript span controls from native code [#682](https://github.com/bugsnag/bugsnag-js-performance/pull/682)
+
 ## [v2.15.0] (2025-08-12)
 
 ### Changed